Add Window type
